+++ common/src/main/java/dev/terminalmc/chatnotify/mixin/accessor/MultilineTextFieldAccessor.java
@@ -23,6 +23,9 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(MultilineTextField.class)
 public interface MultilineTextFieldAccessor {
+    @Accessor
+    int getCursor();
+    
     @Accessor("cursor") 
     void setCursor(int cursor);